# Support any dictionary types JsonExtensionData