--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverter.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverter.cs
@@ -79,5 +79,8 @@ namespace System.Text.Json.Serialization
         internal virtual bool ConstructorIsParameterized => false;
 
         internal ConstructorInfo? ConstructorInfo { get; set; }
+
+        internal virtual object? ConvertToDictionary(ref ReadStack state, object value, JsonSerializerOptions options)
+            => value;
     }
 }