--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverterFactory.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverterFactory.cs
@@ -46,10 +46,7 @@ namespace System.Text.Json.Serialization
             throw new InvalidOperationException();
         }
 
-        internal override JsonParameterInfo CreateJsonParameterInfo()
-        {
-            throw new InvalidOperationException();
-        }
+        internal sealed override object CreateDictionaryOfStringT() => throw new InvalidOperationException();
 
         internal sealed override Type? ElementType => null;