This changes the way that ExtensionData is gathered and assigned so that properties are added and assigned before finishing parsing an object.

This works for the following dictionary types
- `Dictionary<,>`
- `SortedDictionary<,>`
- `IDictionary<,>`
- `IReadOnlyDictionary<,>`
- `ImmutableDictionary<,>`
- `ImmutableSortedDictionary<,>`
- `IImmutableDictionary<,>`
- `ConcurrentDictionary<,>`

closes #31645