@Romanx 2020-04-30
@layomia I had a look at your suggestions and made a commit merging them along with rebasing all the changes on top of master. I also made some replies on your comments about some cases that are already covered by other tests

I tried running the performance tests using master as a baseline (without any specific data extension tests added) and got the following using a 2% threshold as suggested in the doc:

summary:
better: 7, geomean: 1.042
worse: 16, geomean: 1.035
total diff: 23

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromStr |      1.14 |           614.24 |           698.81 |         |
| System.Text.Json.Serialization.Tests.WriteJson<ArrayList>.SerializeObjectPropert |      1.04 |         14160.18 |         14768.43 |         |
| System.Text.Json.Serialization.Tests.ReadJson<MyEventsListerViewModel>.Deseriali |      1.04 |        452811.41 |        471336.04 |         |
| System.Text.Json.Serialization.Tests.WriteJson<LoginViewModel>.SerializeToStream |      1.04 |           550.35 |           571.60 |         |
| System.Text.Json.Serialization.Tests.ReadJson<LoginViewModel>.DeserializeFromUtf |      1.03 |           560.35 |           577.89 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Int32>.SerializeToUtf8Bytes       |      1.03 |           175.85 |           180.79 |         |
| System.Text.Json.Serialization.Tests.WriteJson<BinaryData>.SerializeToString     |      1.03 |           761.69 |           782.26 |         |
| System.Text.Json.Serialization.Tests.WriteJson<BinaryData>.SerializeObjectProper |      1.03 |           997.93 |          1024.76 |         |
| System.Text.Json.Serialization.Tests.ReadJson<HashSet<String>>.DeserializeFromSt |      1.03 |         18433.70 |         18926.40 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Location>.SerializeToString       |      1.03 |          1080.45 |          1109.17 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ArrayList>.DeserializeFromString   |      1.03 |         59883.01 |         61418.66 |         |
| System.Text.Json.Serialization.Tests.ReadJson<BinaryData>.DeserializeFromStream  |      1.02 |          1108.24 |          1135.76 |         |
| System.Text.Json.Serialization.Tests.ReadJson<Location>.DeserializeFromStream    |      1.02 |          1946.39 |          1994.25 |         |
| System.Text.Json.Serialization.Tests.ReadJson<ImmutableDictionary<String, String |      1.02 |         54959.14 |         56253.62 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Hashtable>.SerializeToUtf8Bytes   |      1.02 |         21127.02 |         21617.07 |         |
| System.Text.Json.Serialization.Tests.ReadJson<SimpleStructWithProperties>.Deseri |      1.02 |           702.34 |           718.30 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Serialization.Tests.WriteJson<SimpleStructWithProperties>.Seria |      1.07 |           611.65 |           571.14 |         |
| System.Text.Json.Serialization.Tests.WriteJson<BinaryData>.SerializeToStream     |      1.05 |           634.36 |           603.74 |         |
| System.Text.Json.Serialization.Tests.WriteJson<IndexViewModel>.SerializeToStream |      1.05 |         22086.72 |         21041.37 |         |
| System.Text.Json.Serialization.Tests.WriteJson<SimpleStructWithProperties>.Seria |      1.04 |           362.26 |           349.30 |         |
| System.Text.Json.Serialization.Tests.WriteJson<SimpleStructWithProperties>.Seria |      1.04 |           436.05 |           420.82 |         |
| System.Text.Json.Serialization.Tests.WriteJson<Dictionary<String, String>>.Seria |      1.03 |         13726.78 |         13355.25 |         |
| System.Text.Json.Serialization.Tests.WriteJson<IndexViewModel>.SerializeToString |      1.02 |         23519.59 |         22983.97 |         |