@layomia 2020-05-01
@Romanx, thanks for benchmarking and sharing these numbers.
The regressions here are concerning. I'll try to repro & take another pass at review for this PR.