@Romanx 2020-05-01
Thanks very much, I'm not sure if it's my machine but I tried to keep a stable baseline during the tests and think I ran them correctly