@layomia 2020-06-17
@Romanx, thank you for all the work and investigation you put into this PR. Unfortunately, the perf risk and complexity is too large at this stage of the release, given that the scenario being enabled is relatively minor given [other JSON feature requests ](https://github.com/dotnet/runtime/projects/25). We'll be willing to consider this in the future.