@Romanx 2020-06-17
@layomia no worries at all, was glad to help at all.